% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calculate_factore_score}
\alias{calculate_factore_score}
\title{Calculate factor scores from a model object}
\usage{
calculate_factore_score(model, save_scores = FALSE)
}
\arguments{
\item{model}{an MIRT model object}

\item{save_scores}{if TRUE, scores will be saved under inst/scores, file name: scores_t1.rds}
}
\description{
Calculate factor scores from a model object
}
