% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures.R
\name{create_figure_7.1}
\alias{create_figure_7.1}
\title{Create figure 7.1}
\usage{
create_figure_7.1(save_file = FALSE)
}
\arguments{
\item{save_file}{if TRUE, a pdf file generated under the working directory. File name: Abbildung7.1.pdf}
}
\value{
base plot of factor loadings
}
\description{
Create figure 7.1
}
