% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures.R
\name{create_figure_7.2}
\alias{create_figure_7.2}
\title{Create figure 7.2}
\usage{
create_figure_7.2(save_file = FALSE)
}
\arguments{
\item{save_file}{if TRUE, a pdf file is generated under. File name: Abbildung7.2.pdf}
}
\value{
base plot of factor loadings
}
\description{
Create figure 7.2
}
