% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures.R
\name{create_figure_7.3}
\alias{create_figure_7.3}
\title{Create figure 7.3}
\usage{
create_figure_7.3(save_file = FALSE)
}
\arguments{
\item{save_file}{if TRUE, a pdf file is generated. File name: Abbildung7.3.pdf}
}
\description{
Create figure 7.3
}
