% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures.R
\name{create_figure_8.1}
\alias{create_figure_8.1}
\title{Create figure 8.1}
\usage{
create_figure_8.1(save_file)
}
\arguments{
\item{save_file}{if TRUE, a pdf file is generated. File name: Abbildung8.1.pdf}
}
\description{
Create figure 8.1
}
