% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_fit_index}
\alias{get_fit_index}
\title{Get fit index from an mirt model object}
\usage{
get_fit_index(x, model_name = "model")
}
\arguments{
\item{x}{an object return by the mirt::mirt() function}
}
\value{
a data frame with all fit index
}
\description{
Get fit index from an mirt model object
}
