% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{reorder_items}
\alias{reorder_items}
\title{Reorder items according to the reports in the manuscript}
\usage{
reorder_items(df)
}
\arguments{
\item{df}{a data frame with row.names which are the item ids}
}
\description{
Reorder items according to the reports in the manuscript
}
