% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{report_item_fit}
\alias{report_item_fit}
\title{Report item fit index from an mirt model object}
\usage{
report_item_fit(model)
}
\arguments{
\item{model}{an MIRT model object returned by mirt::mirt()}
}
\description{
Report item fit index from an mirt model object
}
