% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{run_bifactor_model_full_sample}
\alias{run_bifactor_model_full_sample}
\title{Run bifactor model with whole sample}
\usage{
run_bifactor_model_full_sample(save_model = FALSE)
}
\arguments{
\item{save_model}{if TRUE, then the results are saved under inst/models. File name: model_bifactor_full_sample.rds}
}
\value{
a list of model objects returned by mirt::bfactor()
}
\description{
Run bifactor model with whole sample
}
