% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{run_confirmatory_mirt_22items}
\alias{run_confirmatory_mirt_22items}
\title{Run model confirmatory MIRT with 22 items}
\usage{
run_confirmatory_mirt_22items(save_model = FALSE)
}
\arguments{
\item{save_model}{if TRUE, then the results are saved under inst/models. File name: models_confirmatory.rds}
}
\value{
a list of model objects returned by mirt::mirt() or mirt::bfactor()
}
\description{
Run model confirmatory MIRT with 22 items
}
