% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{run_exploratory_mirt_22items}
\alias{run_exploratory_mirt_22items}
\title{Run model with 22 items (IM06 deleted from the model)}
\usage{
run_exploratory_mirt_22items(save_model = FALSE)
}
\arguments{
\item{save_model}{if TRUE, then the results are saved under inst/models. File name: models_22items.rds}
}
\value{
a list of model objects returned by mirt::mirt()
}
\description{
Run model with 22 items (IM06 deleted from the model)
}
