% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{run_exploratory_mirt_23items}
\alias{run_exploratory_mirt_23items}
\title{Run model with 23 items}
\usage{
run_exploratory_mirt_23items(save_model = FALSE)
}
\arguments{
\item{save_model}{if TRUE, then the results are saved under inst/models. File name: models_23items.rds}
}
\value{
a list of model objects returned by mirt::mirt()
}
\description{
Run model with 23 items
}
