% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{report_loadings}
\alias{report_loadings}
\title{Report factor loadings from an mirt model object}
\usage{
report_loadings(model)
}
\arguments{
\item{model}{an MIRT model object returned by mirt::mirt()}
}
\value{
a data frame with factor loadings & h2. Loadings under .30 are surpressed. Character strings are reported (instead of numeric values)
}
\description{
Report factor loadings from an mirt model object
}
